# application/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/benchmark": {
            "get": {
                "description": "Simulate a benchmark with configurable delay or max_delay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "benchmark"
                ],
                "summary": "Run benchmark",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fixed delay in ms",
                        "name": "delay",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 500,
                        "description": "Max random delay in ms",
                        "name": "max_delay",
                        "in": "query"
                    },
                    {
                        "description": "Override delay/max_delay via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.BenchmarkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "benchmark latency response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Simulate a benchmark with configurable delay or max_delay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "benchmark"
                ],
                "summary": "Run benchmark",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fixed delay in ms",
                        "name": "delay",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 500,
                        "description": "Max random delay in ms",
                        "name": "max_delay",
                        "in": "query"
                    },
                    {
                        "description": "Override delay/max_delay via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.BenchmarkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "benchmark latency response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Simulate a benchmark with configurable delay or max_delay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "benchmark"
                ],
                "summary": "Run benchmark",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fixed delay in ms",
                        "name": "delay",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 500,
                        "description": "Max random delay in ms",
                        "name": "max_delay",
                        "in": "query"
                    },
                    {
                        "description": "Override delay/max_delay via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.BenchmarkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "benchmark latency response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "description": "Simulate a benchmark with configurable delay or max_delay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "benchmark"
                ],
                "summary": "Run benchmark",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Fixed delay in ms",
                        "name": "delay",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 500,
                        "description": "Max random delay in ms",
                        "name": "max_delay",
                        "in": "query"
                    },
                    {
                        "description": "Override delay/max_delay via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.BenchmarkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "benchmark latency response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/kpi/availability": {
            "get": {
                "description": "Simulate service availability with configurable success rate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kpi"
                ],
                "summary": "Check availability",
                "parameters": [
                    {
                        "type": "number",
                        "default": 99.9,
                        "description": "Success rate (0–100)",
                        "name": "success_rate",
                        "in": "query"
                    },
                    {
                        "description": "Override success_rate via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.AvailabilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "available response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "unavailable response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Simulate service availability with configurable success rate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kpi"
                ],
                "summary": "Check availability",
                "parameters": [
                    {
                        "type": "number",
                        "default": 99.9,
                        "description": "Success rate (0–100)",
                        "name": "success_rate",
                        "in": "query"
                    },
                    {
                        "description": "Override success_rate via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.AvailabilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "available response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "unavailable response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Simulate service availability with configurable success rate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kpi"
                ],
                "summary": "Check availability",
                "parameters": [
                    {
                        "type": "number",
                        "default": 99.9,
                        "description": "Success rate (0–100)",
                        "name": "success_rate",
                        "in": "query"
                    },
                    {
                        "description": "Override success_rate via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.AvailabilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "available response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "unavailable response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "description": "Simulate service availability with configurable success rate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kpi"
                ],
                "summary": "Check availability",
                "parameters": [
                    {
                        "type": "number",
                        "default": 99.9,
                        "description": "Success rate (0–100)",
                        "name": "success_rate",
                        "in": "query"
                    },
                    {
                        "description": "Override success_rate via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.AvailabilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "available response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "unavailable response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/kpi/performance": {
            "get": {
                "description": "Simulate latency with configurable max delay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kpi"
                ],
                "summary": "Measure performance",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 500,
                        "description": "Max delay in ms",
                        "name": "max_delay",
                        "in": "query"
                    },
                    {
                        "description": "Override max_delay via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.PerformanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "latency response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Simulate latency with configurable max delay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kpi"
                ],
                "summary": "Measure performance",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 500,
                        "description": "Max delay in ms",
                        "name": "max_delay",
                        "in": "query"
                    },
                    {
                        "description": "Override max_delay via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.PerformanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "latency response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Simulate latency with configurable max delay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kpi"
                ],
                "summary": "Measure performance",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 500,
                        "description": "Max delay in ms",
                        "name": "max_delay",
                        "in": "query"
                    },
                    {
                        "description": "Override max_delay via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.PerformanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "latency response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "description": "Simulate latency with configurable max delay",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kpi"
                ],
                "summary": "Measure performance",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 500,
                        "description": "Max delay in ms",
                        "name": "max_delay",
                        "in": "query"
                    },
                    {
                        "description": "Override max_delay via JSON",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.PerformanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "latency response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.AvailabilityRequest": {
            "type": "object",
            "properties": {
                "success_rate": {
                    "type": "number"
                }
            }
        },
        "main.BenchmarkRequest": {
            "type": "object",
            "properties": {
                "delay": {
                    "description": "Pointer to distinguish between 0 and not provided",
                    "type": "integer"
                },
                "max_delay": {
                    "type": "integer"
                }
            }
        },
...
        "main.PerformanceRequest": {
            "type": "object",
            "properties": {
                "max_delay": {
                    "type": "integer"
                }
            }
        }
    }
}